+++ src/vma/proto/dst_entry_tcp.cpp
@@ -363,7 +363,7 @@ void dst_entry_tcp::configure_headers()
     dst_entry::configure_headers();
 }
 
-ssize_t dst_entry_tcp::pass_buff_to_neigh(const iovec *p_iov, size_t sz_iov, uint16_t packet_id)
+ssize_t dst_entry_tcp::pass_buff_to_neigh(const iovec *p_iov, size_t sz_iov, uint32_t packet_id)
 {
     NOT_IN_USE(packet_id);
     m_header_neigh->init();
